Fix compile errors with Visual Studio 2010 32-bit for Issue #224 - Added a (double) cast as input to sqrt in eigenverb_test.cc:300 - Added a static_cast<long double> as input to floor seq_augment.h:64 - Added a static_cast<long double> as input to log10 in beam_pattern_line.cc:95 - Added a static_cast<double> as input to sqrt in beam_pattern_test.cc:507 - Changed size_t to int: in quadtree_test.cc:35, 37
Note that usml_test did not pass without errors.  More work still to do.

+++ types/seq_augment.h
@@ -61,7 +61,7 @@ protected:
 
         double first = (*origin)[0] ;
         _value = _index_data = first ;
-        size_t half = floor(num_rays/2)+1 ;
+        size_t half = floor((double)num_rays/2)+1 ;
         double spacing = origin->increment(0) / half ;
         _data[0] = first ;
         for(size_t i=1; i<half+1; ++i) {
